chore: move initClient.ts and initReloadServer.ts to /initializers
