+++ packages/hmr/lib/initializers/initClient.ts
@@ -1,6 +1,6 @@
-import { LOCAL_RELOAD_SOCKET_URL } from './constant';
-import MessageInterpreter from './interpreter';
-import { DO_UPDATE, DONE_UPDATE } from './types';
+import { LOCAL_RELOAD_SOCKET_URL } from '../constant';
+import MessageInterpreter from '../interpreter';
+import { DO_UPDATE, DONE_UPDATE } from '../types';
 
 export default function initClient({ id, onUpdate }: { id: string; onUpdate: () => void }) {
   const ws = new WebSocket(LOCAL_RELOAD_SOCKET_URL);
