+++ packages/hmr/lib/initializers/initReloadServer.ts
@@ -1,8 +1,8 @@
 import type { WebSocket } from 'ws';
 import { WebSocketServer } from 'ws';
-import { LOCAL_RELOAD_SOCKET_PORT, LOCAL_RELOAD_SOCKET_URL } from './constant';
-import MessageInterpreter from './interpreter';
-import { BUILD_COMPLETE, DO_UPDATE, DONE_UPDATE } from './types';
+import { LOCAL_RELOAD_SOCKET_PORT, LOCAL_RELOAD_SOCKET_URL } from '../constant';
+import MessageInterpreter from '../interpreter';
+import { BUILD_COMPLETE, DO_UPDATE, DONE_UPDATE } from '../types';
 
 const clientsThatNeedToUpdate: Set<WebSocket> = new Set();
